Changed some of the controllers to adapt to the new database schema.

+++ app/Http/Controllers/UserController.php
@@ -21,7 +21,7 @@ class UserController extends Controller
   public function showProfile()
   {
     $user = User::find(Auth::id());
-    //$this->authorize('show', $user);
+    $this->authorize('show', $user);
 
     return view('pages.profile.user_profile', [
       'user' => $user,
@@ -40,6 +40,7 @@ class UserController extends Controller
   public function updateProfile(Request $request, $id)
   {
     $user = User::find($id);
+    $this->authorize('update', $user);
 
     if(!is_null($request->input('username'))){
       $user->username = $request->input('username');
@@ -69,7 +70,7 @@ class UserController extends Controller
   public function delete()
   {
     $user = User::find(Auth::id());
-    //$this->authorize('delete', $user);
+    $this->authorize('delete', $user);
     $user->delete();
 
     return redirect('/');
